dayatani_chatbot/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
anyio==4.0.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.0
backcall==0.2.0
blinker==1.6.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
dataclasses-json==0.5.14
decorator==5.1.1
Django==3.2
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
duckduckgo-search==3.8.4
executing==1.2.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
langchain==0.0.247
langchainplus-sdk==0.0.20
langsmith==0.0.49
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pgvector==0.1.8
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
azure-cognitiveservices-speech==1.33.0
ffmpeg-python==0.2.0
requests_toolbelt==1.0.0
redis==4.5.5
rq==1.14.1